requirements.txt
flask
flask-restful
pymongo
python-dotenv
flask-jwt-extended
bcrypt
python-jose
gunicorn
dnspython
flasgger
flask-cors
certifi